Fix clippy (#1426)
* Fix clippy

* Fix clippy

+++ tests/crud/updates.rs
@@ -19,7 +19,7 @@ pub async fn test_update_cake(db: &DbConn) {
         price: Set(dec!(10.25)),
         gluten_free: Set(false),
         serial: Set(Uuid::new_v4()),
-        bakery_id: Set(Some(bakery_insert_res.last_insert_id as i32)),
+        bakery_id: Set(Some(bakery_insert_res.last_insert_id)),
         ..Default::default()
     };
